+++ src/main/java/net/sf/jabref/gui/prrv/PrrvAction.java
@@ -1,0 +1,26 @@
+package net.sf.jabref.gui.prrv;
+
+import java.awt.event.ActionEvent;
+
+import javax.swing.Action;
+import javax.swing.Icon;
+
+import javafx.application.Platform;
+
+import net.sf.jabref.gui.actions.MnemonicAwareAction;
+
+public class PrrvAction extends MnemonicAwareAction {
+
+    public PrrvAction(String title, String tooltip, Icon iconFile) {
+        super(iconFile);
+        putValue(Action.NAME, title);
+        putValue(Action.SHORT_DESCRIPTION, tooltip);
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+
+        if(!PrrvDialogView.isStarted)
+        Platform.runLater(() -> new PrrvDialogView().show());
+    }
+}