+++ src/main/java/net/sf/jabref/model/entry/InternalBibtexFields.java
@@ -68,6 +68,8 @@ public class InternalBibtexFields {
     private static final List<String> BIBLATEX_LANGUAGE_FIELDS = Arrays.asList(FieldName.LANGUAGE,
             FieldName.ORIGLANGUAGE);
 
+    private static final List<String> BIBLATEX_CITES_KEY_FIELDS = Arrays.asList(FieldName.CITES, FieldName.ENTRYSET);
+
     private static final List<String> BIBLATEX_MULTI_KEY_FIELDS = Arrays.asList(FieldName.RELATED, FieldName.ENTRYSET);
 
     private static final List<String> VERBATIM_FIELDS = Arrays.asList(FieldName.URL, FieldName.FILE,
@@ -335,6 +337,16 @@ public class InternalBibtexFields {
 
         // Set all fields with multiple key links
         for (String fieldText : BIBLATEX_MULTI_KEY_FIELDS) {
+            BibtexSingleField field = fieldSet.get(fieldText);
+            if (field == null) {
+                field = new BibtexSingleField(fieldText, true, BibtexSingleField.SMALL_W);
+            }
+            field.getFieldProperties().add(FieldProperty.CITES);
+            add(field);
+        }
+
+        // Set all fields with cite key links
+        for (String fieldText : BIBLATEX_CITES_KEY_FIELDS) {
             BibtexSingleField field = fieldSet.get(fieldText);
             if (field == null) {
                 field = new BibtexSingleField(fieldText, true, BibtexSingleField.SMALL_W);